requirements.txt
feedparser
requests
beautifulsoup4
newspaper3k
lxml_html_clean
pymongo
python-dotenv
Flask
flask-cors